src_repo/requirements.txt
lxml
scikit-learn
torch==1.5.1
torchvision==0.6.1
pillow

